backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib
python-multipart
python-dotenv
email-validator
